Separated enums from employee model
