+++ src/app/models/designation.model.ts
@@ -1,0 +1,8 @@
+export enum Designation {
+  Assistant = 'Assistant',
+  DataScientist = 'Data Scientist',
+  Developer = 'Developer',
+  Manager = 'Manager',
+  Tester = 'Tester',
+  Trainee = 'Trainee'
+}
