+++ src/app/models/employee.model.ts
@@ -1,3 +1,8 @@
+import { MaritalStatus } from './marital-status.model';
+import { Grade } from './grade.model';
+import { Designation } from './designation.model';
+import { Gender } from './gender.model';
+
 export interface EmployeeList {
   [key: string]: Employee[];
 }
@@ -21,37 +26,9 @@ export class Employee {
   ) {}
 }
 
-export enum Designation {
-  Assistant = 'Assistant',
-  DataScientist = 'Data Scientist',
-  Developer = 'Developer',
-  Manager = 'Manager',
-  Tester = 'Tester',
-  Trainee = 'Trainee'
-}
 
-export enum Gender {
-  Female = 'Female',
-  Male = 'Male'
-}
 
-export enum Grade {
-  M1 = 'M1',
-  M2 = 'M2',
-  M3 = 'M3',
-  M4 = 'M4',
-  M5 = 'M5',
-  M6 = 'M6',
-  M7 = 'M7'
-}
 
-export enum MaritalStatus {
-  Divorced = 'Divorced',
-  Married = 'Married',
-  Separated = 'Separated',
-  Single = 'Single',
-  Widowed = 'Widowed'
-}
 
 // Converts JSON strings to/from your types
 export class Convert {
