+++ src/app/models/gender.model.ts
@@ -1,0 +1,4 @@
+export enum Gender {
+  Female = 'Female',
+  Male = 'Male'
+}
