+++ src/app/models/grade.model.ts
@@ -1,0 +1,9 @@
+export enum Grade {
+  M1 = 'M1',
+  M2 = 'M2',
+  M3 = 'M3',
+  M4 = 'M4',
+  M5 = 'M5',
+  M6 = 'M6',
+  M7 = 'M7'
+}
