+++ src/app/models/marital-status.model.ts
@@ -1,0 +1,7 @@
+export enum MaritalStatus {
+  Divorced = 'Divorced',
+  Married = 'Married',
+  Separated = 'Separated',
+  Single = 'Single',
+  Widowed = 'Widowed'
+}
